[FEATURE] Add Fluid partial for main navigation
The example provides Fluid sections for the first menu level
and all sub-levels. Necessary classes can easily be added or changed.

A condition also sets 'is-current' classes for pages of type 'shortcut'.

By default, 3 levels are rendered via the MenuProcessor in TypoScript.
All sub-levels in this simple demo will be hidden with CSS, though.